+++ tests/Package/PackageFile/Writer/PackageJsonWriterTest.php
@@ -17,8 +17,8 @@ use Puli\RepositoryManager\Package\PackageFile\PackageFile;
 use Puli\RepositoryManager\Package\PackageFile\RootPackageFile;
 use Puli\RepositoryManager\Package\PackageFile\Writer\PackageJsonWriter;
 use Puli\RepositoryManager\Package\ResourceMapping;
-use Puli\RepositoryManager\Package\TagDefinition;
-use Puli\RepositoryManager\Package\TagMapping;
+use Puli\RepositoryManager\Tag\TagDefinition;
+use Puli\RepositoryManager\Tag\TagMapping;
 use Puli\RepositoryManager\Tests\JsonWriterTestCase;
 use Symfony\Component\Filesystem\Filesystem;